ref: extract agent execution and agent response in different component.
