+++ src/components/agents/AgentDetail.tsx
@@ -1,16 +1,13 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { useParams } from "react-router-dom";
 import {
   Bot,
   Calendar,
-  Mail,
-  Send,
   Copy,
   Loader2,
   AlertCircle,
   CheckCircle,
   Maximize2,
-  X,
 } from "lucide-react";
 import { formatDistanceToNow } from "date-fns";
 
@@ -22,9 +19,6 @@ import {
   CardTitle,
 } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
-import { Textarea } from "@/components/ui/textarea";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Separator } from "@/components/ui/separator";
 import { Alert, AlertDescription } from "@/components/ui/alert";
@@ -34,6 +28,8 @@ import { useAgent } from "@/hooks/useAgents";
 import { useRunAgent } from "@/hooks/useRunAgent";
 import { useAgentStore } from "@/store/useAgentStore";
 import { MarkdownRenderer } from "./MarkdownRenderer";
+import { AgentFullScreenResponseModal } from "./modals/AgentFullScreenResponseModal";
+import { AgentExecutionForm } from "./AgentExecutionForm";
 import { toast } from "sonner";
 import IncubyteLeaf from "../icons/incubyteLeaf";
 
@@ -47,62 +43,23 @@ export const AgentDetail = () => {
   const runAgentMutation = useRunAgent(agentId);
   const { isRunning } = useAgentStore();
 
-  const [userPrompt, setUserPrompt] = useState("");
-  const [email, setEmail] = useState("");
   const [isFullScreenOpen, setIsFullScreenOpen] = useState(false);
   const [currentResult, setCurrentResult] = useState<string | null>(null);
 
-  // Handle escape key to close modal
-  useEffect(() => {
-    const handleEscape = (event: KeyboardEvent) => {
-      if (event.key === "Escape" && isFullScreenOpen) {
-        setIsFullScreenOpen(false);
-      }
-    };
-
-    if (isFullScreenOpen) {
-      document.addEventListener("keydown", handleEscape);
-      // Prevent body scroll when modal is open
-      document.body.style.overflow = "hidden";
-    }
-
-    return () => {
-      document.removeEventListener("keydown", handleEscape);
-      document.body.style.overflow = "unset";
-    };
-  }, [isFullScreenOpen]);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (!userPrompt.trim()) {
-      toast.error("Please enter a prompt");
-      return;
-    }
-
-    if (email.trim()) {
-      // Validate email format
-      const emailPattern = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-      if (!emailPattern.test(email)) {
-        toast.error("Please enter a valid email address");
-        return;
-      }
-    }
+  const handleAgentExecution = async (data: { prompt: string; email: string }) => {
     try {
       const result = await runAgentMutation.mutateAsync({
-        prompt: userPrompt.trim(),
-        user_email: email.trim(),
+        prompt: data.prompt,
+        user_email: data.email,
       });
       // Store result in local state instead of global store
       setCurrentResult(result.response || "Response generated successfully");
-
-      // Clear form on success
-      setUserPrompt("");
     } catch (error) {
       console.error("Error running agent:", error);
       toast.error(
         error instanceof Error ? error.message : "Failed to run agent"
       );
+      throw error; // Re-throw to let form handle it
     }
   };
 
@@ -193,84 +150,11 @@ export const AgentDetail = () => {
         {/* Interaction Panel */}
         <div className="grid grid-cols-1 lg:grid-cols-2 gap-8">
           {/* Input Form */}
-          <Card className="border-gray-200 bg-white">
-            <CardHeader>
-              <CardTitle className="flex items-center gap-2">
-                <Send className="h-5 w-5 text-primary" />
-                Execute Agent
-              </CardTitle>
-              <CardDescription>
-                Enter your prompt and email to interact with this agent
-              </CardDescription>
-            </CardHeader>
-
-            <CardContent>
-              <form onSubmit={handleSubmit} className="space-y-4">
-                <div className="space-y-2">
-                  <Label htmlFor="prompt">Your Prompt</Label>
-                  <Textarea
-                    id="prompt"
-                    placeholder={agentPrompt ?? "Type your prompt here..."}
-                    value={userPrompt}
-                    onChange={(e) => setUserPrompt(e.target.value)}
-                    className="min-h-32 resize-none"
-                    required
-                  />
-                  <div className="flex justify-between text-xs text-gray-500">
-                    <span>{userPrompt.length} characters</span>
-                    <span>Be specific for better results</span>
-                  </div>
-                </div>
-
-                <div className="space-y-2">
-                  <Label htmlFor="email">Email Address</Label>
-                  <div className="relative">
-                    <Mail className="absolute left-3 top-1/2 -translate-y-1/2 h-4 w-4 text-gray-400" />
-                    <Input
-                      id="email"
-                      type="email"
-                      placeholder="your.email@example.com"
-                      value={email}
-                      onChange={(e) => setEmail(e.target.value)}
-                      className="pl-10"
-                    />
-                  </div>
-                </div>
-
-                <div className="flex gap-3 pt-2">
-                  <Button
-                    type="submit"
-                    disabled={isRunning || !userPrompt.trim()}
-                    className="flex-1 bg-primary hover:bg-primary/90"
-                  >
-                    {isRunning ? (
-                      <>
-                        <Loader2 className="h-4 w-4 mr-2 animate-spin" />
-                        Processing...
-                      </>
-                    ) : (
-                      <>
-                        <Send className="h-4 w-4 mr-2" />
-                        Generate Response
-                      </>
-                    )}
-                  </Button>
-
-                  <Button
-                    type="button"
-                    variant="outline"
-                    onClick={() => {
-                      setUserPrompt("");
-                      setEmail("");
-                    }}
-                    disabled={isRunning}
-                  >
-                    Clear
-                  </Button>
-                </div>
-              </form>
-            </CardContent>
-          </Card>
+          <AgentExecutionForm
+            agentPrompt={agentPrompt}
+            isRunning={isRunning}
+            onSubmit={handleAgentExecution}
+          />
 
           {/* Results Panel */}
           <Card className="border-gray-200 bg-white">
@@ -347,57 +231,13 @@ export const AgentDetail = () => {
         </div>
 
         {/* Full Screen Modal */}
-        {isFullScreenOpen && (
-          <div className="fixed inset-0 z-50 flex items-center justify-center">
-            {/* Backdrop */}
-            <div
-              className="absolute inset-0 bg-black/50 backdrop-blur-sm"
-              onClick={() => setIsFullScreenOpen(false)}
-            />
-
-            {/* Modal Content */}
-            <div className="relative w-[95vw] h-[95vh] bg-white rounded-lg shadow-2xl flex flex-col">
-              {/* Header */}
-              <div className="flex items-center justify-between p-6 border-b border-gray-200">
-                <div className="flex items-center gap-2">
-                  <Bot className="h-5 w-5 text-secondary" />
-                  <h2 className="text-lg font-semibold text-gray-900">
-                    Agent Response
-                  </h2>
-                </div>
-                <div className="flex gap-2">
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    onClick={handleCopyResult}
-                    className="border-secondary/20 hover:bg-secondary/5"
-                  >
-                    <Copy className="h-4 w-4 mr-2" />
-                    Copy
-                  </Button>
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    onClick={() => setIsFullScreenOpen(false)}
-                  >
-                    <X className="h-4 w-4" />
-                  </Button>
-                </div>
-              </div>
-
-              {/* Content */}
-              <div className="flex-1 min-h-0 p-6">
-                <ScrollArea className="h-full w-full border rounded-md">
-                  <div className="p-6">
-                    {currentResult && (
-                      <MarkdownRenderer content={currentResult} />
-                    )}
-                  </div>
-                </ScrollArea>
-              </div>
-            </div>
-          </div>
-        )}
+        <AgentFullScreenResponseModal
+          isOpen={isFullScreenOpen}
+          onClose={() => setIsFullScreenOpen(false)}
+          content={currentResult}
+          onCopy={handleCopyResult}
+          title={`Response from ${agent.name}`}
+        />
       </div>
     </div>
   );
