+++ src/components/agents/modals/index.ts
@@ -1,0 +1,1 @@
+export { AgentFullScreenResponseModal } from './AgentFullScreenResponseModal';
